fix(arch): split line-limit instantiation and formatting helpers
